% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeActorObject.R
\name{actor}
\alias{actor}
\title{Create actor Object}
\usage{
actor(countData, reduceTissue = TRUE, tissueList = NULL)
}
\arguments{
\item{countData}{Sample isoform data. First column should be titled "gene_id" and second "feature_id"}

\item{reduceTissue}{TRUE or FALSE indicating whether or not the model should reduce to the top 5 tissues.}

\item{tissueList}{List of tissues that will be used for fitting the model. If left NULL all tissues will be used.}
}
\value{
directorObj - Object to be used for model fitting
}
\description{
To scale and flip contributions from a cross validation multi-omics analysis in order to be align with the contributions from the full analysis.
}
