% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{actorFit}
\alias{actorFit}
\title{Fit model}
\usage{
actorFit(actorObj)
}
\arguments{
\item{actorObj}{actor object from `actor`. Model specifications can be set within the object.}
}
\value{
Object of class `actorFit` which can be used for post model fitting interpretations and analyses.
}
\description{
To scale and flip contributions from a cross validation multi-omics analysis in order to be align with the contributions from the full analysis.
}
