% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availTissues.R
\name{availableTissues}
\alias{availableTissues}
\title{Print the available list of GTEx tissues. If no object is provided, the function provides all available GTEx tissues.}
\usage{
availableTissues(actorObj = NULL)
}
\arguments{
\item{actorObj}{Output from actor model}
}
\value{
A vector of all available tissues to plot
}
\description{
Print the available list of GTEx tissues. If no object is provided, the function provides all available GTEx tissues.
}
