% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isoData}
\alias{isoData}
\title{Isoform data from the motor neuron dataset used as the real dataset from the paper.}
\format{A matrix with 10 columns and 20768 rows.}
\source{
\url{https://trace.ncbi.nlm.nih.gov/Traces/sra/?study=SRP067645}
}
\usage{
isoData
}
\description{
A matrix of isoform expression data.
}
\keyword{datasets}
