% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGenePlots.R
\name{makeGenePlots}
\alias{makeGenePlots}
\title{Create Gene Level Plots}
\usage{
makeGenePlots(actorObj, isoData, geneList, tissueList = NULL,
  outFile = NULL)
}
\arguments{
\item{actorObj}{Output from actor model}

\item{isoData}{Sample data to be included in heatmap. First column must be "gene_id" and second column "feature_id"}

\item{geneList}{List of genes to plot.}

\item{tissueList}{List of tissues to plot. Must be a subset of the tissues used in `actorObj`. If left blank then the columns of `likelihoodSum` from `actorObj` will be used.}

\item{outFile}{PDF file path to write gene plots to. Must end in .pdf.}
}
\description{
To scale and flip contributions from a cross validation multi-omics analysis in order to be align with the contributions from the full analysis.
}
