% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.actorFit}
\alias{plot.actorFit}
\title{Create plots for visualization of output}
\usage{
\method{plot}{actorFit}(x, plotType, addAnnot = FALSE,
  annotMat_Col = NULL, annotMat_Row = NULL, fontSize = 20, ...)
}
\arguments{
\item{x}{Object of class `actorFit`}

\item{plotType}{Specifies the type of plot to be generated
\code{"TissueMem"} - Plots posterior probabilities for the tissue membership  \cr\cr
\code{"ClassMem"} - Plots posterior probabilities for gene class membership \cr\cr
\code{"GeneClass"} - Plots gene class posterior estimates of each tissue.}

\item{addAnnot}{TRUE or FALSE for adding annotation bars to the heatmap.}

\item{annotMat_Col}{Dataframe for annotation bars to be added to the columns of the heatmap}

\item{annotMat_Row}{Dataframe for annotation bars to be added to the rows of the heatmap}

\item{fontSize}{Font size for text in plot}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Create plots for visualization of output
}
