% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availTissues.R
\name{availableTissues}
\alias{availableTissues}
\title{Print the available list of GTEx tissues.}
\usage{
availableTissues()
}
\value{
A vector of all available tissues
}
\description{
Print the available list of GTEx tissues.
}
